native: 64 bit compatibility and Linux/x86_64 support for the native
Initial version to test 64 bit compatibility.

Instead of a separate board, 64 bit for Linux/x86_64 is enabled by setting the environment variable `NATIVE_64BIT=y` and compiling as usual.
While I personally prefer a separate `native64` board, this ensures that all tests run with the same configuration as the 32 bit version.
A separate board would require a major refactoring of many tests, which often have special behavior for the native board,
and I didn't want to increase the size of an already large pull request.

Not currently implemented:
* Architectures other than x86_64 or operating systems other than Linux
    * No FreeBSD support
    * No Aarch support
* Rust support for x86_64

+++ cpu/native/periph/flashpage.c
@@ -55,9 +55,9 @@ void flashpage_write(void *target_addr, const void *data, size_t len)
     assert((uintptr_t)target_addr >= (uintptr_t)_native_flash);
     assert((uintptr_t)target_addr + len <= (uintptr_t)_native_flash + sizeof(_native_flash));
     assert(!(len % FLASHPAGE_WRITE_BLOCK_SIZE));
-    assert(!((unsigned)target_addr % FLASHPAGE_WRITE_BLOCK_ALIGNMENT));
+    assert(!((uintptr_t)target_addr % FLASHPAGE_WRITE_BLOCK_ALIGNMENT));
 
-    DEBUG("%p: write %u bytes\n", target_addr, len);
+    DEBUG("%p: write %zu bytes\n", target_addr, len);
 
     _flash_write(target_addr, data, len);
 }
